requirements.txt
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
Flask==1.1.2
Flask-Mail==0.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
make-response==1
MarkupSafe==1.1.1
marshmallow==3.8.0
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
